Dynamically update cutter, strategy, and postprocessor types

+++ cam_refactor/props/utils.py
@@ -1,12 +1,4 @@
-import importlib
-
 import bpy
-
-modnames = ["camjob"]
-
-globals().update(
-    {modname: importlib.reload(importlib.import_module(f".{modname}", __package__)) for modname in modnames}
-)
 
 PRECISION = 5
 
@@ -18,7 +10,7 @@ def get_propnames(pg: bpy.types.PropertyGroup, use_exclude_propnames=True):
     return sorted({propname for propname in pg.rna_type.properties.keys() if propname not in exclude_propnames})
 
 
-def copy(from_prop: bpy.types.Property, to_prop: bpy.types.Property, depth=0) -> None:
+def copy(context: bpy.types.Context, from_prop: bpy.types.Property, to_prop: bpy.types.Property, depth=0) -> None:
     def noop(*args, **kwargs) -> None:
         pass
 
@@ -29,26 +21,28 @@ def copy(from_prop: bpy.types.Property, to_prop: bpy.types.Property, depth=0) ->
 
             from_subprop = getattr(from_prop, propname)
             if any(isinstance(from_subprop, t) for t in [bpy.types.PropertyGroup, bpy.types.bpy_prop_collection]):
-                copy(from_subprop, getattr(to_prop, propname), depth + 1)
+                copy(context, from_subprop, getattr(to_prop, propname), depth + 1)
             elif hasattr(to_prop, propname):
-                if any(isinstance(from_subprop, t) for t in [bpy.types.Collection, bpy.types.Object]):
+                if propname == "data" and any(
+                    isinstance(from_subprop, t) for t in [bpy.types.Collection, bpy.types.Object]
+                ):
                     from_subprop = from_subprop.copy()
                     link = noop
-                    if isinstance(from_prop, camjob.CAMJob):
-                        link = bpy.context.collection.children.link
-                        bpy.context.scene.cam_job_active_index += 1
+                    if isinstance(from_subprop, bpy.types.Collection):
+                        link = context.collection.children.link
+                        context.scene.cam_job_active_index += 1
                         for obj in from_subprop.objects:
                             from_subprop.objects.unlink(obj)
-                    elif isinstance(from_prop, camjob.operation.Operation):
+                    elif isinstance(from_subprop, bpy.types.Object):
                         from_subprop.data = from_subprop.data.copy()
-                        link = bpy.context.scene.cam_job.data.objects.link
+                        link = context.scene.cam_job.data.objects.link
                     link(from_subprop)
                 setattr(to_prop, propname, from_subprop)
 
     elif isinstance(from_prop, bpy.types.bpy_prop_collection):
         to_prop.clear()
         for from_subprop in from_prop.values():
-            copy(from_subprop, to_prop.add(), depth + 1)
+            copy(context, from_subprop, to_prop.add(), depth + 1)
 
 
 def poll_object_source(strategy: bpy.types.Property, obj: bpy.types.Object) -> bool:
